- move render types - delete unused files - expose event emitter
