--- src/rendering/mesh/shared/sword-puzlle checks
@@ -1,24 +1,0 @@
-// migrate:
-// math-extra
-// astro
-// spine
-// sound
-// MeshShader => MeshMaterial
-// Code Splitting!
-// shader precision   settings.PRECISION_FRAGMENT = PRECISION.LOW;  settings.PRECISION_VERTEX = PRECISION.HIGH;
-// renderer options...
-// accessibility
-// build is slow?? transforming each time?
-// BLUR filter!
-// manifest - could be way smaller! webp / 0.5 etc rather than full string
-// recompiling assets??
-// assets WE LOAD TOO MUCH
-// use manifest inn a better way!
-// Better error handling in assets
-// SPine atlas loader- hacked in the web stuffs
-// migrate z index
-// stroke Style - not working with types
-// graphics add functions to main object
-// Spine need to measure it out so layout works
-// 'fff4e6' should be a valid color - intergare color class!
-// removing the mask on the cheat screen makes some crazy batch stuff happen - looks like i need to check with more than 16 texturesd a bit more robustly!
