+++ src/utils/const.ts
@@ -9,4 +9,9 @@
  *
  * DATA_URI.test('data:image/png;base64,foobar'); // => true
  */
+import EventEmitter from 'eventemitter3';
+
 export const DATA_URI = /^\s*data:(?:([\w-]+)\/([\w+.-]+))?(?:;charset=([\w-]+))?(?:;(base64))?,(.*)/i;
+
+// export the event emitter so we can use it in external modules
+export { EventEmitter };
